backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
python-dotenv==1.0.0
